requirements.txt
boto3
psycopg2-binary
pre-commit
python-dotenv
pandas